+++ Sources/StreamChatSwiftUI/ChatChannel/Composer/MessageComposerView.swift
@@ -345,13 +345,15 @@ public struct ComposerInputView<Factory: ViewFactory>: View, KeyboardReadable {
                    displayInfo.isInstant == true {
                     HStack(spacing: 0) {
                         Image(uiImage: images.smallBolt)
+                            .renderingMode(.template)
+                            .foregroundColor(Color(colors.staticColorText))
                         Text(displayInfo.displayName.uppercased())
                     }
                     .padding(.horizontal, 8)
                     .font(fonts.footnoteBold)
                     .frame(height: 24)
-                    .background(Color.blue)
-                    .foregroundColor(.white)
+                    .background(colors.tintColor)
+                    .foregroundColor(Color(colors.staticColorText))
                     .cornerRadius(16)
                 }