+++ StreamChatSwiftUITests/Tests/ChatChannel/MessageComposerView_Tests.swift
@@ -402,4 +402,41 @@ class MessageComposerView_Tests: StreamChatTestCase {
         // Then
         assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
     }
+
+    func test_composerInputView_command() {
+        let factory = DefaultViewFactory.shared
+        let size = CGSize(width: defaultScreenSize.width, height: 100)
+
+        let view = ComposerInputView(
+            factory: factory,
+            text: .constant(""),
+            selectedRangeLocation: .constant(0),
+            command: .constant(.init(
+                id: .unique,
+                typingSuggestion: .empty,
+                displayInfo: CommandDisplayInfo(
+                    displayName: "Giphy",
+                    icon: Images().commandGiphy,
+                    format: "",
+                    isInstant: true
+                )
+            )),
+            addedAssets: [],
+            addedFileURLs: [],
+            addedCustomAttachments: [],
+            quotedMessage: .constant(nil),
+            cooldownDuration: 0,
+            onCustomAttachmentTap: { _ in },
+            removeAttachmentWithId: { _ in }
+        )
+        .environmentObject(MessageComposerTestUtils.makeComposerViewModel(chatClient: chatClient))
+        .frame(width: size.width, height: size.height)
+
+        AssertSnapshot(view, variants: .onlyUserInterfaceStyles, size: size)
+
+        // Themed
+        streamChat?.appearance.colors.tintColor = .mint
+        streamChat?.appearance.colors.staticColorText = .black
+        AssertSnapshot(view, variants: .onlyUserInterfaceStyles, size: size, suffix: "themed")
+    }
 }